refactor checkAccountOwnership to middleware/near, added this to 2fa routes
