+++ app.js
@@ -37,6 +37,7 @@ const router = new Router();
 const {
     creatorKeyJson,
     withNear,
+    checkAccountOwnership,
 } = require('./middleware/near');
 
 app.use(withNear);
@@ -50,33 +51,14 @@ const {
     sendNewCode,
     verifyCode,
 } = require('./middleware/2fa');
-router.post('/2fa/getAccessKey', getAccessKey);
-router.post('/2fa/init', initCode);
-router.post('/2fa/send', sendNewCode);
-router.post('/2fa/verify', verifyCode);
+router.post('/2fa/getAccessKey', checkAccountOwnership, getAccessKey);
+router.post('/2fa/init', checkAccountOwnership, initCode);
+router.post('/2fa/send', checkAccountOwnership, sendNewCode);
+router.post('/2fa/verify', checkAccountOwnership, verifyCode);
 
 const VALID_BLOCK_AGE = 100;
 
-async function checkAccountOwnership(ctx, next) {
-    const { accountId, blockNumber, blockNumberSignature } = ctx.request.body;
-    if (!accountId || !blockNumber || !blockNumberSignature) {
-        ctx.throw(403, 'You must provide an accountId, blockNumber, and blockNumberSignature');
-    }
 
-    const currentBlock = (await ctx.near.connection.provider.status()).sync_info.latest_block_height;
-    const givenBlock = Number(blockNumber);
-
-    if (givenBlock <= currentBlock - VALID_BLOCK_AGE || givenBlock > currentBlock) {
-        ctx.throw(403, `You must provide a blockNumber within ${VALID_BLOCK_AGE} of the most recent block; provided: ${blockNumber}, current: ${currentBlock}`);
-    }
-
-    const nearAccount = await ctx.near.account(accountId);
-    if (!(await verifySignature(nearAccount, blockNumber, blockNumberSignature))) {
-        ctx.throw(403, `blockNumberSignature did not match a signature of blockNumber=${blockNumber} from accountId=${accountId}`);
-    }
-
-    return await next();
-}
 
 async function checkAccountDoesNotExist(ctx, next) {
     const { accountId } = ctx.request.body;
@@ -110,29 +92,7 @@ const SECURITY_CODE_DIGITS = 6;
 
 const { sendSms } = require('./utils/sms');
 
-const nacl = require('tweetnacl');
-const crypto = require('crypto');
-const bs58 = require('bs58');
-const verifySignature = async (nearAccount, data, signature) => {
-    try {
-        const hash = crypto.createHash('sha256').update(data).digest();
-        const accessKeys = (await nearAccount.getAccessKeys());
-        /********************************
-        @todo what should be the correct filter here
-        ********************************/
-        // .filter(({ access_key: { permission } }) => permission === 'FullAccess' ||
-        //         permission.FunctionCall &&
-        //             permission.FunctionCall.receiver_id === nearAccount.accountId &&
-        //             permission.FunctionCall.method_names.includes('__wallet__metadata'));
-        return accessKeys.some(it => {
-            const publicKey = it.public_key.replace('ed25519:', '');
-            return nacl.sign.detached.verify(hash, Buffer.from(signature, 'base64'), bs58.decode(publicKey));
-        });
-    } catch (e) {
-        console.error(e);
-        return false;
-    }
-};
+
 
 async function recoveryMethodsFor(account) {
     if (!account) return [];
